+++ src/test/java/use_case/createquiz/CohereClientTest.java
@@ -2,7 +2,7 @@ package use_case.createquiz;
 
 import entity.Quiz;
 import entity.QuizQuestion;
-import use_case.create_quiz.CohereAPI;
+import use_case.create_quiz.cohere_interaction.CohereAPI;
 
 public class CohereClientTest {
     public static void main(String[] args) {
